fix: unblock native windows build
